update and rename
added:
  * localized wikipedia sources

fixed:
  * some typos

enhanced:
  * more facts
  * rename for consistent naming and better overview in folder
